% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_by_puuid.R
\name{get_account_by_puuid}
\alias{get_account_by_puuid}
\title{Fetch Riot Account Data by PUUID}
\usage{
get_account_by_puuid(puuid, region, api_key)
}
\arguments{
\item{puuid}{A character string representing the player's PUUID (Player Universally Unique ID).}

\item{region}{A character string representing the player's region (america, asia, europe)}

\item{api_key}{A character string containing your valid Riot API key.}
}
\value{
A list containing the player's PUUID, gameName, and tagLine.
}
\description{
Retrieves a player's Riot Account information (PUUID, gameName, and tagLine)
using the Riot Account-V1 endpoint, specifically by providing their PUUID.
}
\examples{
\dontrun{
# Replace with your actual API key and a valid PUUID
api_key <- "YOUR_API_KEY"
example_puuid <- "YOUR_EXAMPLE_PUUID_HERE"
account_info_by_puuid <- get_account_by_riot_id(puuid = example_puuid, api_key = api_key)
print(account_info_by_puuid$gameName)
}

}
